Fixing nlohmann-json compilation issues (#21)
This PR fixes compilation issues with an older version of nlohmann-json on the Github Actions Ubuntu runner than was available on MacPorts. It also removes `using` directives from header files, and applies autoformatting using `clang-format`. All the tests are now built and run on each commit, not just the parser tests.